requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-filter==23.2
django-tinymce==3.6.1
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
